Updates for compatibility with CDT HEAD

+++ photran/org.eclipse.photran.cdtinterface/src/org/eclipse/photran/internal/cdtinterface/ui/FViewLabelProvider.java
@@ -31,7 +31,7 @@ public class FViewLabelProvider extends CViewLabelProvider
 
     private Image fortranFileImage;
 
-    public FViewLabelProvider(int textFlags, int imageFlags)
+    public FViewLabelProvider(long textFlags, int imageFlags)
     {
         super(textFlags, imageFlags);
     }
